MOBWOT-15486 Added line plot
Former-commit-id: e52e1e03e94b07310e3e95bb3f97e2cece72ea3b

+++ Tools/ResourceEditor/Classes/Qt/DeviceInfo/DeviceInfo/MemProfInfoModel.cpp
@@ -151,7 +151,7 @@ void MemProfInfoModel::setConfig(const DAVA::MMStatConfig* statConfig)
     emit headerDataChanged(Qt::Horizontal, 0, columnCount());
     emit headerDataChanged(Qt::Vertical, 0, rowCount());
 }
-void MemProfInfoModel::forTagStats(std::function<void(int, const TagsStat&)> onStat)
+void MemProfInfoModel::forTagStats(std::function<void(size_t, const TagsStat&)> onStat)
 {
     auto it = timedData.begin();
     while (it != timedData.end())
